(WIP) Ensure PEM symlinks are present
This patch ensures that symlinks used by the PEM installer are present before
installation is attempted. This is necessary for installs that are run locally
using archives that are shared using the /vagrant guest file system. These
archives are extracted using the Ruby Minitar library, for Windows support, and
Minitar doesn't handle symlinks.

Once PEM becomes the default installer, these symlinks may disappear. This
patch can be reverted when this change occurs.